post-intex/comments/likes functionalities

+++ app/controllers/api/likes_controller.rb
@@ -1,29 +1,30 @@
-class Api::LikesController < ApplicationController  
-  def create
-    @like = Like.new(like_params)
-    @like.user_id = current_user.id
+class Api::LikesController < ApplicationController
+    before_action :require_logged_in
 
-    if @like.save 
-            render :show, status: 200
+    def create 
+        @like = Like.new(like_params)
+        @like.user_id = current_user.id 
+
+        if @like.save 
+            render :show
         else
             render json: @like.errors.full_messages, status: 422
         end
-  end
-
-  def destroy
-    @like = Like.where(user_id: current_user.id)
-        .where(post_id: params[:id])[0]
-
-    if @like.destroy 
-        render :show, status: 200
-    else
-        render json: ["can't be processed"], status: 422
     end
-  end
 
-  private
+    def destroy
+        @like = Like.find_by(id: params[:id])
+        if @like && @like.user_id == current_user.id
+            if @like.destroy 
+                render json: {}, status: 200
+            else
+                render json: ["can't be processed"], status: 422
+            end
+        end
+    end
 
-  def like_params
-    params.require(:like).permit(:user_id, :post_id)
-  end
+    private
+    def like_params
+        params.require(:like).permit(:post_id)
+    end
 end
